requirements.txt
mock==2.0.0
nose==1.3.7
requests==2.10.0
python-dateutil==2.5.3
beautifulsoup4==4.5.1
pyproj==1.9.5.1
Shapely==1.5.16
tzwhere==2.3